Add DTOs/services and update controllers
Introduce a large refactor for the sgac-final module: added many Request/Response DTOs, service interfaces (I*), and numerous service implementations under service/impl; updated repositories and entities accordingly. Controllers were refactored to depend on service interfaces; several controller endpoint implementations (Asignatura, Carrera, Convocatoria) were commented out and AuthController was extended with multiple registration endpoints (student, docente, coordinador, decano, ayudante). Removed legacy entity/service/repository classes (PlazoActividad, TipoPlazoActividad and related services). Also updated application.properties. Additionally, small changes in the PFEjemplo demo: Operaciones interface now returns int and Main adjusted to match (including some functional/Consumer/Supplier examples) and CredencialesUsuario updated.

+++ PFEjemplo/src/Operaciones.java
@@ -1,3 +1,3 @@
 public interface Operaciones {
-    float Operacion(int a, int b);
+    int Operacion(int a, int b);
 }
